chore: Move mock db to tests folder
