+++ src/app/submission/test/submission.test.ts
@@ -1,6 +1,6 @@
+import { MockDatabase } from './MockDatabase';
 import * as snsSampleAnonymous from './samples/sns.sample-anonymous.json';
 import * as snsSample from './samples/sns.sample.json';
-import { MockDatabase } from '../Database';
 import { MockFormConnector } from '../FormConnector';
 import { MockStorage } from '../Storage';
 import { Submission } from '../Submission';
